requirements.txt
ipykernel
pandas
numpy
certifi
fastapi
gradio
python-multipart
pillow
opencv-python
torch
torchvision
torchsummary
matplotlib
seaborn
scikit-learn
albumentations
pillow
#onnxruntime
python-dotenv
pymongo

